# `MyPy` and `Pylint` partition inputs via `CoarsenedTarget`